wpt_babel/content-security-policy/securitypolicyviolation/source-file-data-scheme.html

<!DOCTYPE html>
<meta content="script-src 'nonce-abc' data:" http-equiv="Content-Security-Policy"/>
<script nonce="abc" src="/resources/testharness.js">
</script>
<script nonce="abc" src="/resources/testharnessreport.js">
</script>
<script nonce="abc">
 "use strict";

async_test(function (t) {
  var watcher = new EventWatcher(t, document, 'securitypolicyviolation');
  watcher.wait_for('securitypolicyviolation').then(t.step_func_done(function (e) {
    assert_equals(e.blockedURI, "eval");
    assert_equals(e.sourceFile, "data");
    assert_equals(e.lineNumber, 3);
    assert_equals(e.columnNumber, 17);
  }));
  var scriptText = "\n            try {\n                eval(\"assert_unreached('no eval')\");\n            } catch (e) {\n                assert_equals(e.name, 'EvalError');\n            }\n        ";
  var s = document.createElement("script");
  s.src = "data:text/javascript," + encodeURIComponent(scriptText);
  document.head.append(s);
}, "Violations from data:-URL scripts have a sourceFile of 'data'");
</script>
